Add turn management and word handling in game logic
- Introduced TurnStaredPayload for turn events in WebSocket protocol
- Updated GameStruct to use int for CurrentRound and MaxRound
- Implemented word selection and masking functionality
- Enhanced GameScreen to display current word and drawer player

+++ server/internal/domain/game/game.go
@@ -2,8 +2,8 @@ package game
 
 type GameStruct struct {
 	Status                string
-	CurrentRound          int64
-	MaxRound              int64
+	CurrentRound          int
+	MaxRound              int
 	DrawerIndex           int
 	CurrentDrawerPlayerId string
 	CurrentWord           string
